BQ MCP/requirement.txt
google-adk
google-generativeai
python-dotenv
google-cloud-aiplatform
fastapi==0.115.0
matplotlib
uuid